# Faster anagram.go for small inputs