#### Is your solution the most efficient way to solve a problem? Why?

It's a more efficient way for small inputs because maps don't need to be initialized.

`go test -bench .` results:

```
goos: darwin
goarch: amd64
BenchmarkTestAnagrams-8             5000            224418 ns/op
PASS
ok      _/Users/admin/git/practice-go/anagram   1.154s
```

Old results: 
```
goos: darwin
goarch: amd64
BenchmarkTestAnagrams-8             2000            917811 ns/op
PASS
ok      _/Users/admin/git/practice-go/anagram   1.942s
```

#### Have you used any specific algorithm?

Nope

#### What is time and space complexity of your solution?

Time complexity is O(`dictionary_length` * `main_word_length` * `dictionary_word_length`). Space complexity is O(`dictionary_length`)