@plutov 2019-07-13
Nice solution! Can you please do one thing: instead of modifying existing file, create a new solution? This way we can keep both.