@qpwo 2019-07-13
You're right; that's better. I think this is all ready to squash and merge now